requirements.txt
pyrogram
tgcrypto
motor
aiohttp
dnspython
psutil
apscheduler
python-dotenv
